gralloc4: Fix data types used for sizing buffer
The scheme followed is:
1. All the inputs retain their original data types.
2. All the calculated dimensions are in uint64_t.
3. All the sizes are in size_t.
4. All the alignment requirements are in uint32_t.

Bug: 297275730
Test: Boot to home
Test: VtsHalGraphicsMapperV4_0TargetTest
Test: b/297275730#comment15
Test: aion_test

+++ gralloc4/src/core/mali_gralloc_bufferdescriptor.h
@@ -50,8 +50,8 @@ struct buffer_descriptor_t
 	 * Calculated values that will be passed to the allocator in order to
 	 * allocate the buffer.
 	 */
-	uint64_t alloc_sizes[MAX_PLANES];
-	int pixel_stride;
+	size_t alloc_sizes[MAX_PLANES];
+	size_t pixel_stride;
 	uint64_t alloc_format;
 	uint32_t fd_count;
 	uint32_t plane_count;
@@ -85,12 +85,12 @@ struct buffer_descriptor_t
 			"format_type(%u) "
 			"name(%s)"
 			"reserved_size(%" PRIu64 ") "
-			"alloc_sizes(%" PRIu64 ", %" PRIu64 ", %" PRIu64 ")"
-			"pixel_stride(%d) alloc_format(0x%" PRIx64 ") fd_count(%d) "
+			"alloc_sizes(%zu, %zu, %zu)"
+			"pixel_stride(%zu) alloc_format(0x%" PRIx64 ") fd_count(%d) "
 			"plane_count(%u) "
-			"plane[0](offset %" PRId64 ", idx %u, size %" PRIu64 " byte_stride %u, wh %u %u)"
-			"plane[1](offset %" PRId64 ", idx %u, size %" PRIu64 " byte_stride %u, wh %u %u)"
-			"plane[2](offset %" PRId64 ", idx %u, size %" PRIu64 " byte_stride %u, wh %u %u)"
+			"plane[0](offset %" PRId64 ", idx %u, size %zu byte_stride %zu, wh %" PRIu64 " %" PRIu64 ")"
+			"plane[1](offset %" PRId64 ", idx %u, size %zu byte_stride %zu, wh %" PRIu64 " %" PRIu64 ")"
+			"plane[2](offset %" PRId64 ", idx %u, size %zu byte_stride %zu, wh %" PRIu64 " %" PRIu64 ")"
 			"\n",
 			str.c_str(),
 			width, height,
